Hacking puzzle mini game
The hacking puzzle mini game is completed for the prototype. Basic assets have been made in the form of a straight "pipe" and angled "pipe". 
Only one mini game can be added to a single scene due to a GameObject.Find call . This will be changed for later version. To edit the puzzle simply change values on the TileEdtor script which is attached to all tiles for the min game.